fix: move restart api call to correct folder
