+++ client/src/api/user/index.js
@@ -1,3 +1,2 @@
-// eslint-disable-next-line import/prefer-default-export
 export { default as getUser } from './get';
-export { default as restartUser } from './post';
+export { default as restartUser } from './restart';
